+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 
 # Changelog
 
+## New version
+
+### New features
+
+- In the `Wizaplace\Catalog\AttributeVariant` class, added the "position" and "seoKeywords" fields which are returned by the API endpoint.
+
 ## 1.135.3
 
 ### New features