+++ CHANGELOG.md
@@ -7,6 +7,19 @@
 
 - Added `Wizaplace\SDK\Discussion\DiscussionService::startDiscussionWithCustomer`
 
+## New version
+
+### New features
+
+- In the `Wizaplace\Catalog\AttributeVariant` class, added the "position" and "seoKeywords" fields which are returned by the API endpoint.
+
+## 1.135.3
+
+### New features
+
+- Added `Wizaplace\SDK\Vendor\Order\OrderChild`
+- Added `Wizaplace\SDK\Vendor\Order\OrderService::getOrderChildren`
+
 ## 1.134.0
 
 - Added `Wizaplace\SDK\Order\OrderService::getOrderShipments`